test: fold the Groonga unit tests into the existing suite
Applies the review feedback from #3794 to this PR.

- Lift the async mock-client fixture pair into `conftest.py` too. Only the
  sync pair was shared before, so `test_groonga_document_store.py` kept its
  own async pair and the new tests added a third, near-identical one.
- Rename the shared fixtures to `mock_groonga_client` / `mock_groonga_store`
  (plus the async variants). `mock_supabase_client` only ever mocked the
  Groonga store's client, which is misleading now that it sits in a conftest
  shared with the pgvector and bucket-downloader tests.
- Drop `test_groonga_document_store_unit.py` and merge it into
  `test_groonga_document_store.py`. The `_unit` suffix carried no meaning —
  both files were unit tests, the integration ones live in
  `test_groonga_integration.py` — and it split adjacent code paths across
  files: in-memory filter matching was tested in both.
- Split the resulting file, and `TestRetriever`, into focused test classes
  (`TestInit`, `TestWarmUp`, `TestSerde`, `TestRun`, ...).

One duplicate goes away: `test_write_documents_empty` and
`test_writing_nothing_is_a_no_op` asserted the same thing.
